clk: qcom: Import GDSC flags from Linux
In Linux driver, depending on which GDSC flags are set, different
status register is checked when powering it on/off. And on top of
that, different bit inside that register is tested.

As a first step towards adding behaviour variation, import GDSC
flags from Linux driver. Even though only POLL_CFG_GDSCR is the
important one here, add all flags in order to make it easier to
compare which flags are needed for which GDSC, making porting
process easier.

Signed-off-by: Alexey Minnekhanov <alexeymin@minlexx.ru>