clk: qcom: Treat different GDSCs differently
In Linux driver, depending on which GDSC flags are set, different
status register is checked when powering it on/off. And on top of
that, different bit inside that register is tested. Currently qcom
clk implementation in U-Boot lacks this functionality, so adjust
logic when checking GDSC status to match one in Linux driver.

This fixes problems on e.g. MSM8998 and SDM660's with USB_30_GDSC
to be forever stuck during power on, since unlike e.g SDM845 one
it doesn't have POLL_CFG_GDSCR flag (only VOTABLE).

With this change all GDSCs that have POLL_CFG_GDSCR in their flags
keep old behaviour (same as before this patch).

Signed-off-by: Alexey Minnekhanov <alexeymin@minlexx.ru>